Extract shared test fixture for patched run_kanpan
Problem: Five test functions in cli_test.py duplicated the same setup
pattern (creating called_with list and monkeypatching run_kanpan).
Fix: Created a patched_run_kanpan fixture in conftest.py that handles
the monkeypatching and yields the captured call list. Updated all five
tests to use the fixture, removing the duplicated setup code.
